// bank/src/types/db.d.ts
export type Database = {
// ...
  public: {
    Tables: {
      accounts: {
        Row: {
          account_number: string
          account_type: Database["public"]["Enums"]["account_type_enum"]
          balance: number
          created_at: string
          id: number
          name: string
          pin_code: string | null
          status: Database["public"]["Enums"]["account_status_enum"]
          user_id: number
        }
        Insert: {
          account_number: string
          account_type: Database["public"]["Enums"]["account_type_enum"]
          balance?: number
          created_at?: string
          id?: number
          name: string
          pin_code?: string | null
          status?: Database["public"]["Enums"]["account_status_enum"]
          user_id: number
        }
        Update: {
          account_number?: string
          account_type?: Database["public"]["Enums"]["account_type_enum"]
          balance?: number
          created_at?: string
          id?: number
          name?: string
          pin_code?: string | null
          status?: Database["public"]["Enums"]["account_status_enum"]
          user_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "accounts_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      payment_schedule: {
        Row: {
          account_id: number
          amount: number
          created_at: string
          day_of_month: number | null
          day_of_week: Database["public"]["Enums"]["day_enum"] | null
          day_of_year: string | null
          end_date: string
          frequency: Database["public"]["Enums"]["schedule_frequency_enum"]
          id: number
          start_date: string
          status: Database["public"]["Enums"]["schedule_status_enum"]
        }
        Insert: {
          account_id: number
          amount: number
          created_at?: string
          day_of_month?: number | null
          day_of_week?: Database["public"]["Enums"]["day_enum"] | null
          day_of_year?: string | null
          end_date: string
          frequency: Database["public"]["Enums"]["schedule_frequency_enum"]
          id?: number
          start_date: string
          status?: Database["public"]["Enums"]["schedule_status_enum"]
        }
        Update: {
          account_id?: number
          amount?: number
          created_at?: string
          day_of_month?: number | null
          day_of_week?: Database["public"]["Enums"]["day_enum"] | null
          day_of_year?: string | null
          end_date?: string
          frequency?: Database["public"]["Enums"]["schedule_frequency_enum"]
          id?: number
          start_date?: string
          status?: Database["public"]["Enums"]["schedule_status_enum"]
        }
        Relationships: [
          {
            foreignKeyName: "payment_schedule_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never